duplicated code removed and made a encapsulated trait for it.
