+++ src/RenderableTrait.php
@@ -1,0 +1,53 @@
+<?php
+
+
+namespace Revolution\LaminasForm;
+
+
+use Illuminate\Container\Container;
+use Illuminate\Contracts\Container\BindingResolutionException;
+use Illuminate\Support\HtmlString;
+use Laminas\View\Renderer\RendererInterface;
+
+trait RenderableTrait
+{
+    /**
+     * @return RendererInterface
+     *
+     * @throws BindingResolutionException
+     */
+    protected function getRenderer(): RendererInterface
+    {
+        return Container::getInstance()->make(RendererInterface::class);
+    }
+
+    /**
+     * @param string $helper
+     *
+     * @return HtmlString
+     *
+     * @throws BindingResolutionException
+     * @throws \BadMethodCallException
+     */
+    public function render(string $helper = 'form'): HtmlString
+    {
+        $renderer = $this->getRenderer();
+        return new HtmlString($renderer->$helper($this));
+    }
+
+    /**
+     * @param string $method
+     * @param array $arguments
+     *
+     * @return mixed
+     *
+     * @throws BindingResolutionException
+     * @throws \BadMethodCallException
+     */
+    public function __call(string $method, array $arguments = [])
+    {
+        $renderer = $this->getRenderer();
+
+        return call_user_func_array([$renderer, $method], $arguments);
+    }
+}
